requirements.txt
discord.py
requests
python-dotenv
qrcode
pillow